extract ConversationList to file
